backend/requirements.txt
fastapi
uvicorn
python-multipart
numpy
opencv-python
mediapipe
gTTS
SpeechRecognition
pydantic
pydantic-settings
python-dotenv
requests
aiofiles
httpx
groq
langdetect